Moved content to assets;added resume;Updated color scheme;removed projects section temporarily
