+++ src/Index.css
@@ -1,20 +1,19 @@
 :root {
-  --light-text: rgb(3, 37, 48);
-  --accent-color: rgb(3, 37, 48); /*rgb(22, 220, 223);*/
-  --base-bg-color: rgb(243, 243, 223);
-  --base-hover-color: rgb(16, 51, 97);
-  --base-hover-color-light: rgb(219, 219, 202);
+  --base-text-color: #0c1844;
+  --accent-color: #c80036;
+  --base-bg-color: #fff5e1;
+  --base-hover-color: #f2e6cd;
 }
 @supports (font-variation-settings: normal) {
   :root {
-    font-family: InterVariable, sans-serif;
+    font-family: "Poppins", sans-serif;
     font-feature-settings: "liga" 1, "calt" 1;
   }
 }
 
 body {
   background-color: var(--base-bg-color);
-  color: var(--light-text);
+  color: var(--base-text-color);
 }
 
 @media only screen and (min-width: 800px) {
