+++ src/components/intro/IntroStyles.css
@@ -1,7 +1,7 @@
 .intro-container {
   display: flex;
   flex-basis: 30%;
-  padding-top: 20px;
+  margin-top: 20px;
   flex-direction: column;
   align-items: center;
   justify-content: center;
@@ -11,18 +11,17 @@
   display: flex;
   flex-direction: column;
   width: 60%;
-  h1,
-  h3 {
-    margin: 0;
-    padding: 10px 0px;
-  }
 
   h1 {
+    margin-top: 10px;
+    margin-bottom: 0px;
+    color: var(--accent-color);
     font-size: x-large;
   }
 
   h3 {
-    font-size: medium;
+    margin: 0;
+    font-size: large;
   }
 }
 
@@ -57,4 +56,20 @@
   li:hover {
     transform: translateY(-8px);
   }
+}
+
+.intro-url {
+  text-decoration: none;
+  transition: transform 0.5s ease;
+  color: inherit;
+  font-weight: 600;
+  display: inline-block;
+  line-height: 1.5em;
+  margin-bottom: 10px;
+}
+
+.intro-url:hover {
+  transform: translateX(10px);
+  text-decoration: underline dashed var(--base-text-color) 2px;
+  text-underline-offset: 6px;
 }
