+++ src/components/wrappers/wrappersStyles.css
@@ -20,13 +20,17 @@
 
 .exp-desc {
   flex-basis: 80%;
+  h2 {
+    color: var(--accent-color);
+  }
+
   h3 {
-    font-size:medium;
+    font-size: medium;
     font-weight: 600;
   }
 
   p {
-    font-size: small;
+    font-size: smaller;
   }
 
   a {
@@ -39,7 +43,7 @@
 
   a:hover {
     transform: translateX(10px);
-    text-decoration: underline dashed var(--light-text) 2px;
+    text-decoration: underline dashed var(--base-text-color) 2px;
     text-underline-offset: 6px;
   }
 }
@@ -54,10 +58,9 @@
 }
 
 .exp-skill {
-  /* background-color: var(--accent-color); */
   border: 2px solid var(--accent-color);
-  /* color: var(--base-bg-color); */
   font-size: small;
+  font-weight: 500;
   padding: 5px 5px;
   margin: 2px;
   border-radius: 10px;
