+++ src/icons/CloseIcon.tsx
@@ -9,7 +9,6 @@ Copyright 2024 Fonticons, Inc.
 export default function CloseIcon(props: IconProps) {
   return (
     <>
-      {/* <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 384 512"> */}
       <svg
         className={props.isDark ? "icon-dark" : "icon-light"}
         xmlns="http://www.w3.org/2000/svg"
