+++ src/icons/iconStyles.css
@@ -1,6 +1,6 @@
 .icon-dark {
   path {
-    fill: var(--light-text);
+    fill: var(--base-text-color);
   }
 }
